Safari-609.1.11.1/LayoutTests/fast/canvas/webgl/index-validation-with-subsequent-draws-expected.txt

CONSOLE MESSAGE: line 50: WebGL: INVALID_OPERATION: drawElements: unable to simulate vertexAttrib0 array
CONSOLE MESSAGE: line 56: WebGL: INVALID_OPERATION: drawElements: unable to simulate vertexAttrib0 array
